api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
redis==5.0.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
python-dotenv==1.0.1
